Refactor area selection and editing as hooks and components
